Bug causing a No Shutdown Action
This exception halts the whole Shutdown gui initialization, causing a failure to apply the shutdown action.

Cause: If the ExtendedShutdownDelay value is the default value of 0, it causes an exception when setting the Grace Timer interval (interval can't be 0).

Solution (Not Complete): Moved the setting of the Grace Timer interval only if Grace Interval is used.

(This error will happen as well if user sets the DelayToShutdown parameter as 0)