Added User selectable Frequency source
With My UPS, only an output frequency reading is available. I presume they did in this manner as the output frequency is the same as the input when in line mode (only routed to output), and if it is using the battery, only the inverter output frequency is available.

So I have added a combo box to select the Input/Output as the source. This causes the text above the Frequency guage to change, and the correct reading to be retrieved from the UPS Network. The new field was added to the registry Parameters as well.

I added 2 strings to the settings resource to help with the translation of the guage heading label. I am sure there is a better way...

If possible, please check the translations. I resized the input controls on the calibration tab in the preferences form, so there ara quite a few changes in the language file.